Rename `pretokenizers` -> `pre_tokenizers`
Align with the naming convention in the tokenizers crate.
