+++ rten-text/src/lib.rs
@@ -10,7 +10,7 @@
 
 pub mod models;
 pub mod normalizer;
-pub mod pretokenizers;
+pub mod pre_tokenizers;
 pub mod tokenizers;
 
 mod split;
